CM-Co-Occurrence-AVVP_paddle/requirements.txt
moviepy==1.0.3
munch==2.5.0
numpy==1.20.3
pandas==1.3.2
Pillow==9.1.1
pretrainedmodels==0.7.4
scikit_learn==1.1.1
seaborn==0.11.2
thop==0.0.31-2005241907
paddlepaddle-gpu==1.7.1
paddlepaddle-gpuvision==0.8.2
tqdm==4.62.2
